requirements.txt
twine==3.5.0
click==8.0.3
nuitka==0.6.17.4
requests==2.25.1
pytest==6.2.5
pyyaml==6.0